tdh provisioning script update link creations

+++ bin/03-tdh-provision.sh
@@ -8,8 +8,8 @@ if [ -z "$hadooproot" ]; then
     hadooproot="/opt/TDH"
 fi
 
-if [ -e "/etc/hadoop" ]; then
-    if [ -l "/etc/hadoop" ]; then
+if [ -e "$etchadoop" ]; then
+    if [ -l "$etchadoop" ]; then
         ( sudo rm $etchadoop )
     else
         echo "Error ''$etchadoop' exists and is not a link"
@@ -17,14 +17,19 @@ if [ -e "/etc/hadoop" ]; then
     fi
 fi
 
-( sudo ln -s /opt/TDH/etc /etc/hadoop )
+( sudo ln -s ${hadooproot}/etc $etchadoop )
 
-if [ -e "/opt/hadoop" ]; then
-    if [ -l "/opt/hadoop" ]; then
-        ( sudo rm /opt/hadoop; ln -s $hadooproot /opt/hadoop )
-    else
-        echo "/opt/hadoop exists and is not a link"
+if [ "$hadooproot" == "/opt/hadoop" ]; then
+    echo "HADOOP_ROOT set to $hadooproot"
+else
+    if [ -e "/opt/hadoop" ]; then
+        if [ -l "/opt/hadoop" ]; then
+            ( sudo rm /opt/hadoop; ln -s $hadooproot /opt/hadoop )
+            echo "HADOOP_ROOT of ''$hadooproot' linked to /opt/hadoop"
+        else
+            echo "/opt/hadoop exists and is not a link"
+        fi
     fi
 fi
 
-exit 0    
+exit 0
